Deactivation of hidden cubes now works properly
Hidden cubes can't be deactivate on first pass(loop) since the array is still being filled and we can't perform bound checks (this was already present before this commit)
But they can't be deactivated in second pass either, since bound checks are dependent on active state of cubes.
Marking them in second pass and deactivating in third fixes the problem.